requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
markdown2==2.4.10
google-generativeai==0.8.3
gunicorn==21.2.0
supabase==2.3.4
postgrest==0.13.2 